Restructure to allow both borrowed and owned values live together
