+++ src/value/borrowed/from.rs
@@ -1,24 +1,40 @@
 use super::{MaybeBorrowedString, Value};
+use crate::numberparse::Number;
 
 impl<'a> From<&'a str> for MaybeBorrowedString<'a> {
+    #[inline]
     fn from(v: &'a str) -> Self {
         MaybeBorrowedString::B(v)
     }
 }
 
 impl<'a> From<String> for MaybeBorrowedString<'a> {
+    #[inline]
     fn from(v: String) -> Self {
         MaybeBorrowedString::O(v)
     }
 }
 
+impl<'a> From<Number> for Value<'a> {
+    #[inline]
+    fn from(n: Number) -> Self {
+        match n {
+            Number::F64(n) => Value::F64(n),
+            Number::I64(n) => Value::I64(n),
+        }
+    }
+}
+
+/********* str_ **********/
 impl<'a> From<&'a str> for Value<'a> {
+    #[inline]
     fn from(s: &'a str) -> Self {
         Value::String(MaybeBorrowedString::B(s))
     }
 }
 
 impl<'a> From<String> for Value<'a> {
+    #[inline]
     fn from(s: String) -> Self {
         Value::String(MaybeBorrowedString::O(s))
     }
@@ -26,6 +42,7 @@ impl<'a> From<String> for Value<'a> {
 
 /********* atoms **********/
 impl<'a> From<bool> for Value<'a> {
+    #[inline]
     fn from(b: bool) -> Self {
         Value::Bool(b)
     }
@@ -33,48 +50,56 @@ impl<'a> From<bool> for Value<'a> {
 
 /********* i_ **********/
 impl<'a> From<i8> for Value<'a> {
+    #[inline]
     fn from(i: i8) -> Self {
         Value::I64(i as i64)
     }
 }
 
 impl<'a> From<i16> for Value<'a> {
+    #[inline]
     fn from(i: i16) -> Self {
         Value::I64(i as i64)
     }
 }
 
 impl<'a> From<i32> for Value<'a> {
+    #[inline]
     fn from(i: i32) -> Self {
         Value::I64(i as i64)
     }
 }
 
 impl<'a> From<i64> for Value<'a> {
+    #[inline]
     fn from(i: i64) -> Self {
         Value::I64(i as i64)
     }
 }
 
 impl<'a> From<&i8> for Value<'a> {
+    #[inline]
     fn from(i: &i8) -> Self {
         Value::I64(*i as i64)
     }
 }
 
 impl<'a> From<&i16> for Value<'a> {
+    #[inline]
     fn from(i: &i16) -> Self {
         Value::I64(*i as i64)
     }
 }
 
 impl<'a> From<&i32> for Value<'a> {
+    #[inline]
     fn from(i: &i32) -> Self {
         Value::I64(*i as i64)
     }
 }
 
 impl<'a> From<&i64> for Value<'a> {
+    #[inline]
     fn from(i: &i64) -> Self {
         Value::I64(*i as i64)
     }
@@ -82,24 +107,28 @@ impl<'a> From<&i64> for Value<'a> {
 
 /********* f_ **********/
 impl<'a> From<f32> for Value<'a> {
+    #[inline]
     fn from(f: f32) -> Self {
         Value::F64(f as f64)
     }
 }
 
 impl<'a> From<f64> for Value<'a> {
+    #[inline]
     fn from(f: f64) -> Self {
         Value::F64(f as f64)
     }
 }
 
 impl<'a> From<&f32> for Value<'a> {
+    #[inline]
     fn from(f: &f32) -> Self {
         Value::F64(*f as f64)
     }
 }
 
 impl<'a> From<&f64> for Value<'a> {
+    #[inline]
     fn from(f: &f64) -> Self {
         Value::F64(*f as f64)
     }
