+++ src/value/owned/de.rs
@@ -1,9 +1,12 @@
 use super::{Map, MaybeBorrowedString, Value};
-use crate::{stry, Error, Result};
-use serde::de::{self, DeserializeSeed, MapAccess, SeqAccess, Unexpected, Visitor};
+use crate::{stry, Error};
+use serde::de::{
+    self, Deserialize, DeserializeSeed, Deserializer, MapAccess, SeqAccess, Unexpected, Visitor,
+};
 use serde::forward_to_deserialize_any;
 use serde_ext::de::IntoDeserializer;
 use std::borrow::Cow;
+use std::fmt;
 
 impl<'de> de::Deserializer<'de> for Value {
     type Error = Error;
@@ -11,7 +14,7 @@ impl<'de> de::Deserializer<'de> for Value {
     // Look at the input data to decide what Serde data model type to
     // deserialize as. Not all data formats are able to support this operation.
     // Formats that support `deserialize_any` are known as self-describing.
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value>
+    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Error>
     where
         V: Visitor<'de>,
     {
@@ -33,7 +36,7 @@ impl<'de> de::Deserializer<'de> for Value {
     }
 }
 
-fn visit_array<'de, V>(array: Vec<Value>, visitor: V) -> Result<V::Value>
+fn visit_array<'de, V>(array: Vec<Value>, visitor: V) -> Result<V::Value, Error>
 where
     V: Visitor<'de>,
 {
@@ -51,7 +54,7 @@ where
     }
 }
 
-fn visit_object<'de, V>(object: Map, visitor: V) -> Result<V::Value>
+fn visit_object<'de, V>(object: Map, visitor: V) -> Result<V::Value, Error>
 where
     V: Visitor<'de>,
 {
@@ -84,7 +87,7 @@ impl<'de> serde::Deserializer<'de> for SeqDeserializer {
     type Error = Error;
 
     #[inline]
-    fn deserialize_any<V>(mut self, visitor: V) -> Result<V::Value>
+    fn deserialize_any<V>(mut self, visitor: V) -> Result<V::Value, Self::Error>
     where
         V: Visitor<'de>,
     {
@@ -115,7 +118,7 @@ impl<'de> serde::Deserializer<'de> for SeqDeserializer {
 impl<'de> SeqAccess<'de> for SeqDeserializer {
     type Error = Error;
 
-    fn next_element_seed<T>(&mut self, seed: T) -> Result<Option<T::Value>>
+    fn next_element_seed<T>(&mut self, seed: T) -> Result<Option<T::Value>, Self::Error>
     where
         T: DeserializeSeed<'de>,
     {
@@ -150,7 +153,7 @@ impl MapDeserializer {
 impl<'de> MapAccess<'de> for MapDeserializer {
     type Error = Error;
 
-    fn next_key_seed<T>(&mut self, seed: T) -> Result<Option<T::Value>>
+    fn next_key_seed<T>(&mut self, seed: T) -> Result<Option<T::Value>, Self::Error>
     where
         T: DeserializeSeed<'de>,
     {
@@ -166,7 +169,7 @@ impl<'de> MapAccess<'de> for MapDeserializer {
         }
     }
 
-    fn next_value_seed<T>(&mut self, seed: T) -> Result<T::Value>
+    fn next_value_seed<T>(&mut self, seed: T) -> Result<T::Value, Self::Error>
     where
         T: DeserializeSeed<'de>,
     {
@@ -188,7 +191,7 @@ impl<'de> serde::Deserializer<'de> for MapDeserializer {
     type Error = Error;
 
     #[inline]
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value>
+    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
     where
         V: Visitor<'de>,
     {
@@ -208,7 +211,7 @@ struct MapKeyDeserializer<'de> {
 
 macro_rules! deserialize_integer_key {
     ($method:ident => $visit:ident) => {
-        fn $method<V>(self, visitor: V) -> Result<V::Value>
+        fn $method<V>(self, visitor: V) -> Result<V::Value, Self::Error>
         where
             V: Visitor<'de>,
         {
@@ -223,7 +226,7 @@ macro_rules! deserialize_integer_key {
 impl<'de> serde::Deserializer<'de> for MapKeyDeserializer<'de> {
     type Error = Error;
 
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value>
+    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
     where
         V: Visitor<'de>,
     {
@@ -240,7 +243,7 @@ impl<'de> serde::Deserializer<'de> for MapKeyDeserializer<'de> {
     deserialize_integer_key!(deserialize_u64 => visit_u64);
 
     #[inline]
-    fn deserialize_option<V>(self, visitor: V) -> Result<V::Value>
+    fn deserialize_option<V>(self, visitor: V) -> Result<V::Value, Self::Error>
     where
         V: Visitor<'de>,
     {
@@ -249,7 +252,11 @@ impl<'de> serde::Deserializer<'de> for MapKeyDeserializer<'de> {
     }
 
     #[inline]
-    fn deserialize_newtype_struct<V>(self, _name: &'static str, visitor: V) -> Result<V::Value>
+    fn deserialize_newtype_struct<V>(
+        self,
+        _name: &'static str,
+        visitor: V,
+    ) -> Result<V::Value, Self::Error>
     where
         V: Visitor<'de>,
     {
@@ -261,7 +268,7 @@ impl<'de> serde::Deserializer<'de> for MapKeyDeserializer<'de> {
         name: &'static str,
         variants: &'static [&'static str],
         visitor: V,
-    ) -> Result<V::Value>
+    ) -> Result<V::Value, Self::Error>
     where
         V: Visitor<'de>,
     {
@@ -289,7 +296,7 @@ impl<'de> BorrowedCowStrDeserializer<'de> {
 impl<'de> de::Deserializer<'de> for BorrowedCowStrDeserializer<'de> {
     type Error = Error;
 
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value>
+    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
     where
         V: de::Visitor<'de>,
     {
@@ -304,7 +311,7 @@ impl<'de> de::Deserializer<'de> for BorrowedCowStrDeserializer<'de> {
         _name: &str,
         _variants: &'static [&'static str],
         visitor: V,
-    ) -> Result<V::Value>
+    ) -> Result<V::Value, Self::Error>
     where
         V: de::Visitor<'de>,
     {
@@ -322,7 +329,7 @@ impl<'de> de::EnumAccess<'de> for BorrowedCowStrDeserializer<'de> {
     type Error = Error;
     type Variant = UnitOnly;
 
-    fn variant_seed<T>(self, seed: T) -> Result<(T::Value, Self::Variant)>
+    fn variant_seed<T>(self, seed: T) -> Result<(T::Value, Self::Variant), Self::Error>
     where
         T: de::DeserializeSeed<'de>,
     {
@@ -336,11 +343,11 @@ struct UnitOnly;
 impl<'de> de::VariantAccess<'de> for UnitOnly {
     type Error = Error;
 
-    fn unit_variant(self) -> Result<()> {
+    fn unit_variant(self) -> Result<(), Self::Error> {
         Ok(())
     }
 
-    fn newtype_variant_seed<T>(self, _seed: T) -> Result<T::Value>
+    fn newtype_variant_seed<T>(self, _seed: T) -> Result<T::Value, Self::Error>
     where
         T: de::DeserializeSeed<'de>,
     {
@@ -350,7 +357,7 @@ impl<'de> de::VariantAccess<'de> for UnitOnly {
         ))
     }
 
-    fn tuple_variant<V>(self, _len: usize, _visitor: V) -> Result<V::Value>
+    fn tuple_variant<V>(self, _len: usize, _visitor: V) -> Result<V::Value, Self::Error>
     where
         V: de::Visitor<'de>,
     {
@@ -360,7 +367,11 @@ impl<'de> de::VariantAccess<'de> for UnitOnly {
         ))
     }
 
-    fn struct_variant<V>(self, _fields: &'static [&'static str], _visitor: V) -> Result<V::Value>
+    fn struct_variant<V>(
+        self,
+        _fields: &'static [&'static str],
+        _visitor: V,
+    ) -> Result<V::Value, Self::Error>
     where
         V: de::Visitor<'de>,
     {
@@ -368,5 +379,240 @@ impl<'de> de::VariantAccess<'de> for UnitOnly {
             Unexpected::UnitVariant,
             &"struct variant",
         ))
+    }
+}
+
+impl<'de> Deserialize<'de> for Value {
+    fn deserialize<D>(deserializer: D) -> Result<Value, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_any(ValueVisitor)
+    }
+}
+
+struct ValueVisitor;
+
+impl<'de> Visitor<'de> for ValueVisitor {
+    type Value = Value;
+
+    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter.write_str("an JSONesque value")
+    }
+
+    /****************** unit ******************/
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_unit<E>(self) -> Result<Self::Value, E> {
+        Ok(Value::Null)
+    }
+
+    /****************** bool ******************/
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_bool<E>(self, value: bool) -> Result<Self::Value, E> {
+        Ok(Value::Bool(value))
+    }
+
+    /****************** Option ******************/
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_none<E>(self) -> Result<Self::Value, E> {
+        Ok(Value::Null)
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_some<D>(self, deserializer: D) -> Result<Self::Value, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_any(self)
+    }
+
+    /****************** enum ******************/
+    /*
+    fn visit_enum<A>(self, data: A) -> Result<Self::Value, A::Error> where
+        A: EnumAccess<'de>,
+    {
+    }
+     */
+
+    /****************** i64 ******************/
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_i8<E>(self, value: i8) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value as i64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_i16<E>(self, value: i16) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value as i64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_i32<E>(self, value: i32) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value as i64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_i64<E>(self, value: i64) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value))
+    }
+
+    /****************** u64 ******************/
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_u8<E>(self, value: u8) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value as i64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_u16<E>(self, value: u16) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value as i64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_u32<E>(self, value: u32) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::I64(value as i64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_u64<E>(self, value: u64) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        use std::i64;
+        if value <= i64::MAX as u64 {
+            Ok(Value::I64(value as i64))
+        } else {
+            Err(E::custom(format!("Integer out of range: {}", value)))
+        }
+    }
+
+    /****************** f64 ******************/
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_f32<E>(self, value: f32) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::F64(value as f64))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_f64<E>(self, value: f64) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::F64(value))
+    }
+
+    /****************** stringy stuff ******************/
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_char<E>(self, _value: char) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        unimplemented!()
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_borrowed_str<E>(self, value: &'de str) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::from(value))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::String(MaybeBorrowedString::O(value.to_owned())))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_string<E>(self, value: String) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::String(MaybeBorrowedString::O(value)))
+    }
+
+    /****************** byte stuff ******************/
+
+    /*
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_borrowed_bytes<E>(self, value: &'de [u8]) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+        Ok(Value::String(value))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_str<E>(self, value: &[u8]) -> Result<Self::Value, E>
+    where
+    'a: 'de
+        E: de::Error,
+    {
+      Ok(Value::String(value))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_string<E>(self, value: Vec<u8>) -> Result<Self::Value, E>
+    where
+        E: de::Error,
+    {
+      Ok(Value::String(&value))
+    }
+     */
+    /****************** nexted stuff ******************/
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
+    where
+        A: MapAccess<'de>,
+    {
+        let size = map.size_hint().unwrap_or_default();
+
+        let mut m = Map::with_capacity(size);
+        while let Some(k) = map.next_key()? {
+            let v = map.next_value()?;
+            m.insert(k, v);
+        }
+        Ok(Value::Object(m))
+    }
+
+    #[cfg_attr(not(feature = "no-inline"), inline)]
+    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
+    where
+        A: SeqAccess<'de>,
+    {
+        let size = seq.size_hint().unwrap_or_default();
+
+        let mut v = Vec::with_capacity(size);
+        while let Some(e) = seq.next_element()? {
+            v.push(e);
+        }
+        Ok(Value::Array(v))
     }
 }
